requirements.txt
langchain==0.3.19
langchain-community==0.3.18
fastapi>=0.115.0
uvicorn>=0.29.0
websockets>=12.0
streamlit>=1.37.0
websocket-client>=1.7.0
python-dotenv>=1.0.1
pydantic>=2.7.1
